requirements.txt
python-telegram-bot[job-queue]>=21.0
aiohttp>=3.9
aiosqlite>=0.19
python-dotenv>=1.0